Add rh_update_routes, an improved version of rh_add_interface_routes.
The biggest improvement is you no longer have to specify the interface
you want the route on. Instead you specify a regular expression of the
ip that would be on that interface and the specific interface is found.

Second major improvement is the direct use of ip route del/add to update
currently running system instead of restarting all networking.

Also a small change in the declaration of the routes that is less repetive
and requires less typing.

Take care in crafting your ip regular expression especially if you are not
using a standard A, B or C class network.